Merged with Holger's commit from Smalltalkhub. Use strategy design pattern.
Original  comment:

Make the version generator pluggable 

Currently the UUIDGenerator is used to get a random number seed as the version. On Unix platforms this will read from /dev/urandom.  Make the VersionNumber generation pluggable so a different scheme can be used.

Right now and for using libcrypto making this class based and not instance based could be an option. In the future the VersionGenerator could get the object to make different decisions.